requirements.txt
mcp>=1.0.0
python-pptx>=1.0.0
lxml>=5.0.0
pillow>=10.0.0
